refactor: split 8 oversized hooks files into focused modules
Split all hooks crate files exceeding the 400-line limit into
module directories with focused sub-modules:

- session_start.rs (1432 lines) → 5 files (mod, context_loaders, blarify, migration, tests)
- stop/reflection.rs (897) → 3 files (mod, prompt, conversation)
- stop/power_steering.rs (856) → 4 files (mod, analysis, decision, state)
- post_tool_use.rs (764) → 6 files (mod, workflow, validation, metrics, tests, launcher)
- user_prompt.rs (603) → 4 files (mod, preferences, memory, tests)
- pre_tool_use/cwd.rs (512) → 2 files (mod, tests)
- session_stop.rs (467) → 4 files (mod, transcript, git, tests)
- pre_compact.rs (428) → 4 files (mod, export, request, tests)

8 monolithic files → 32 focused modules. All 45 hooks files under
400 lines (largest: 390). 155/155 tests pass, zero regressions.
Public API unchanged.
